+++ tests/Fixtures/Translatable/TranslatableEntityTranslation.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Cloudstek\DoctrineBehaviour\Tests\Fixtures\Translatable;
+namespace mdeboer\DoctrineBehaviour\Tests\Fixtures\Translatable;
 
-use Cloudstek\DoctrineBehaviour\Tests\Fixtures\AbstractEntity;
-use Cloudstek\DoctrineBehaviour\TranslationInterface;
-use Cloudstek\DoctrineBehaviour\TranslationTrait;
+use mdeboer\DoctrineBehaviour\Tests\Fixtures\AbstractEntity;
+use mdeboer\DoctrineBehaviour\TranslationInterface;
+use mdeboer\DoctrineBehaviour\TranslationTrait;
 use Doctrine\ORM\Mapping\Entity;
 
 #[Entity]