+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -6,37 +7,66 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [2.1.0] - 2023-05-24
+## [3.0.0] - 2023-07-30
+
+This release is API compatible with v2.1.0, nothing has changed but the namespace, package name and the repository
+where it is maintained. It is still maintained by me.
+
 ### Changed
+
+- Updated namespace from `Cloudstek\DoctrineBehaviour\` to `mdeboer\DoctrineBehaviour`
+- Update package name from `cloudstek/doctrine-behaviour` to `mdeboer/doctrine-behaviour`
+
+## [2.1.0] - 2023-05-24
+
+### Changed
+
 - Allow DateTimeInterface in setters instead of just DateTime and DateTimeImmutable
 - Update to PHPUnit 10.1
 
 ## [2.0.2] - 2022-06-29
+
 ### Added
+
 - More documentation and examples
 - Changelog
 
 ### Fixed
+
 - Translation not being properly removed with setTranslations().
 
 ## [2.0.1] - 2022-06-29
+
 ### Added
+
 - Translatable behaviour
 - More tests
 - Github action to automatically run tests
 
 ## [2.0.0] - 2022-06-15
+
 ### Changed
+
 - Timestampable columns createdAt and updatedAt are no longer nullable.
 
 ## [1.0.1] - 2022-05-18
+
 ### Added
+
 - Timestampable subscriber to automatically add entity listener to timestampable entities.
 
-[Unreleased]: https://github.com/Cloudstek/doctrine-behaviour/compare/v2.1.0...develop
+[Unreleased]: https://github.com/mdeboer/doctrine-behaviour/compare/v3.0.0...develop
+
+[3.0.0]: https://github.com/mdeboer/doctrine-behaviour/compare/v2.1.0...v3.0.0
+
 [2.1.0]: https://github.com/Cloudstek/doctrine-behaviour/compare/v2.0.2...v2.1.0
+
 [2.0.2]: https://github.com/Cloudstek/doctrine-behaviour/compare/v2.0.1...v2.0.2
+
 [2.0.1]: https://github.com/Cloudstek/doctrine-behaviour/compare/v2.0.0...v2.0.1
+
 [2.0.0]: https://github.com/Cloudstek/doctrine-behaviour/compare/v1.0.1...v2.0.0
+
 [1.0.1]: https://github.com/Cloudstek/doctrine-behaviour/compare/v1.0.0...v1.0.1
+
 [1.0.0]: https://github.com/Cloudstek/doctrine-behaviour/releases/tag/v1.0.0